chore: fold charge-rate edge tests
